+++ lib/src/main/java/com/otaliastudios/transcoder/sink/DataSink.java
@@ -5,8 +5,8 @@ import android.media.MediaFormat;
 
 import androidx.annotation.NonNull;
 
-import com.otaliastudios.transcoder.engine.TrackStatus;
-import com.otaliastudios.transcoder.engine.TrackType;
+import com.otaliastudios.transcoder.common.TrackStatus;
+import com.otaliastudios.transcoder.common.TrackType;
 import com.otaliastudios.transcoder.transcode.TrackTranscoder;
 
 import java.nio.ByteBuffer;
