+++ lib/src/main/java/com/otaliastudios/transcoder/sink/DefaultDataSinkChecks.java
@@ -17,7 +17,7 @@ package com.otaliastudios.transcoder.sink;
 
 import android.media.MediaFormat;
 
-import com.otaliastudios.transcoder.engine.TrackType;
+import com.otaliastudios.transcoder.common.TrackType;
 import com.otaliastudios.transcoder.internal.Logger;
 import com.otaliastudios.transcoder.internal.MediaFormatConstants;
 import com.otaliastudios.transcoder.internal.AvcCsdUtils;
