+++ lib/src/main/java/com/otaliastudios/transcoder/source/DataSourceWrapper.java
@@ -6,7 +6,7 @@ import android.media.MediaFormat;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
-import com.otaliastudios.transcoder.engine.TrackType;
+import com.otaliastudios.transcoder.common.TrackType;
 
 /**
  * A {@link DataSource} wrapper that simply delegates all methods to the
