+++ lib/src/main/java/com/otaliastudios/transcoder/source/FilePathDataSource.java
@@ -3,10 +3,8 @@ package com.otaliastudios.transcoder.source;
 import android.media.MediaExtractor;
 import android.media.MediaMetadataRetriever;
 
-import com.otaliastudios.transcoder.engine.TrackType;
 import com.otaliastudios.transcoder.internal.Logger;
 
-import java.io.FileDescriptor;
 import java.io.FileInputStream;
 import java.io.IOException;
