+++ lib/src/main/java/com/otaliastudios/transcoder/transcode/BaseTrackTranscoder.java
@@ -7,7 +7,7 @@ import android.media.MediaFormat;
 import androidx.annotation.CallSuper;
 import androidx.annotation.NonNull;
 
-import com.otaliastudios.transcoder.engine.TrackType;
+import com.otaliastudios.transcoder.common.TrackType;
 import com.otaliastudios.transcoder.internal.MediaCodecBuffers;
 import com.otaliastudios.transcoder.sink.DataSink;
 import com.otaliastudios.transcoder.source.DataSource;
