+++ lib/src/main/java/com/otaliastudios/transcoder/validator/Validator.java
@@ -2,7 +2,7 @@ package com.otaliastudios.transcoder.validator;
 
 import androidx.annotation.NonNull;
 
-import com.otaliastudios.transcoder.engine.TrackStatus;
+import com.otaliastudios.transcoder.common.TrackStatus;
 import com.otaliastudios.transcoder.strategy.TrackStrategy;
 
 /**
